Fix nested main block on trust details page
As we are working with a side nav, the main emmet tags need to be declared after the side-nav html, but this is already declared in the govuk page template outside of the 'content' block.

Have updated all layouts to use the main block, and moved the alpha banner into the header block as beforeContent is now unavailable.