Reduce `govuk-summary-list` overrides
Create v2 of GOV.UK Summary List style override class, which reduces the number of changes, and keeps it more inline with the original component.

We are still using an override for width on larger screens, so that the description for a property does not wrap at 30%. The width of this column will be that of the largest row.