Fix partial text search showing all results
Fallback option for search will not work with the enhance select option for accessible-autocomplete component, so have changed it to a pure javaScript implementation. If a user selects an option we are storing the trust uid to a hidden input, which is sent as a search parameter.

Refactored trust-search.njk to make it reusable for header and search page implementations. Have also extracted script to keep execution to app script block.